Added hack for multi-gpu fermion force
In multi-gpu mode, I now compute the link-unitarisation fermion-force contribution
over the full extended volume. This may well be a very bad thing to do because
we only need the force contributions in the interior points.
On the other hand, depending on how we handle inter-gpu communication,
it may be more efficient to update the force in the exterior region
since it may reduce the need for communication, maybe.
Need to discuss this with everyone.