Introduced min tolerance in multi-mass solver.
Previously, the mixed-precision multi-mass solver used
the same target residual for the low-precision inner solve
and the  polishing. This doesn't really make sense, and it's
likely that in a typical solve the mixed-precision solver is
doing more work than it needs to. I've fixed this by enforcing
a minimum lower bound on the target relative residual; 1e-6 for
single-precision, and 1e-4 for half precision. I don't yet know if
these settings are optimal.