Freed up some memory in unitarize_link_test.cpp
Unfortunately, many of the functions used in the HISQ routines
allocate memory that is never freed. This isn't a real memory leak
since memory is not allocated repeatedly. However, it does result
in a whole bunch of warnings. This is poor code design, which will
be addressed in the future.