Small fix for APE smearing
A call to the APE smearing funciton would create a new gauge field called gaugeSmeared, copy gaugePrecise there, and perform the APE steps, saving the final result in gaugeSmeared and leaving gaugePrecise untouched. A subsequent call to the APE smearing function will start smearing again the fields in gaugeSmeared (which are already smeared), instead of reloading the gaugePrecise field. I think this is not what we want, although this can be the object of discussion.