Use newer hipblas API from v3 onwards
The hipblasComplexDouble type is the old hipblas v1 interface.  It was
deprecated in v2 (ROCm 6.x), but still usable, and in v3 (ROCm 7.x) it
disappears completely in favour of hipComplexDouble.

The hipblas function interfaces followed suit.  QUDA never bothered trying to
use the newer v2 interface, but it has no choice for v3, so now QUDA understands
the new interface for new enough hipblas libraries and does so with the
intention of not breaking existing users that, for a variety of reasons, may be
using much older ROCm software stacks.

The documentation for the deprecated interface is here:
https://rocm.docs.amd.com/projects/hipBLAS/en/docs-6.4.1/reference/hipblas-api-functions.html#hipblas-v2-and-deprecations