Hardcode 64 wavefront size in the HIP backend
The __GFX9__ macros (and friends) are only defined in the device pass.  Since
the warp_size function is used on the host side, the #else branch was returning
something different than what was returned on the device.

We need to be able to handle devices with different wavefront sizes, and we
need to be able to expose those to the host.  The usual mechanism for host
exposure is to probe the HIP runtime for the current device's warpSize
property, but this approach isn't a constant expression.

In the future, we'll need an approach that either a) doesn't rely on
constexpr-ness; or b) or can rely on constexpr-ness but simply isn't used on
the host.